+++ src/main/java/upc/edu/pe/apileadyourway/serviceinterfaces/IInsightsService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface IInsightsService {
     //List<BicisMenosAlquiladasDTO> obtenerBicicletasMenosAlquiladas(int idSuministrador);
-    List<ReseniaBajaDTO> obtenerReseniaMasBaja(int idSuministrador);
+    List<ReseniaBajaDTO> obtenerReseniaMasBaja(Long idSuministrador);
     //List<ConteoTipoBicicletaDTO> contarAlquileresPorTipo(int idSuministrador);
 }
