+++ src/main/java/upc/edu/pe/apileadyourway/serviceinterfaces/IReporteService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface IReporteService {
     void insertarReporte(Reporte reporte);
     void editarReporte(Reporte reporte);
-    public List<Reporte> listarReporteCliente(int idCliente);
+    public List<Reporte> listarReporteCliente(Long idCliente);
     public Reporte listaID(int id);
     public void eliminarReporte(int id);
 }
